+++ src/types/geometrycollection.rs
@@ -15,15 +15,14 @@
 use std::fmt;
 use std::str::FromStr;
 use tokenizer::{PeekableTokens, Token};
-use FromTokens;
-use Geometry;
+use {FromTokens, Geometry, WktFloat};
 
-#[derive(Clone, Default)]
-pub struct GeometryCollection<T: num_traits::Float>(pub Vec<Geometry<T>>);
+#[derive(Clone, Debug, Default)]
+pub struct GeometryCollection<T: WktFloat>(pub Vec<Geometry<T>>);
 
 impl<T> GeometryCollection<T>
 where
-    T: num_traits::Float,
+    T: WktFloat,
 {
     pub fn as_item(self) -> Geometry<T> {
         Geometry::GeometryCollection(self)
@@ -32,7 +31,7 @@ where
 
 impl<T> fmt::Display for GeometryCollection<T>
 where
-    T: num_traits::Float + fmt::Display,
+    T: WktFloat + fmt::Display,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         if self.0.is_empty() {
@@ -52,7 +51,7 @@ where
 
 impl<T> FromTokens<T> for GeometryCollection<T>
 where
-    T: num_traits::Float + FromStr + Default,
+    T: WktFloat + FromStr + Default,
 {
     fn from_tokens(tokens: &mut PeekableTokens<T>) -> Result<Self, &'static str> {
         let mut items = Vec::new();