+++ src/types/multipoint.rs
@@ -16,15 +16,14 @@ use std::fmt;
 use std::str::FromStr;
 use tokenizer::PeekableTokens;
 use types::point::Point;
-use FromTokens;
-use Geometry;
+use {FromTokens, Geometry, WktFloat};
 
-#[derive(Clone, Default)]
-pub struct MultiPoint<T: num_traits::Float>(pub Vec<Point<T>>);
+#[derive(Clone, Debug, Default)]
+pub struct MultiPoint<T: WktFloat>(pub Vec<Point<T>>);
 
 impl<T> MultiPoint<T>
 where
-    T: num_traits::Float,
+    T: WktFloat,
 {
     pub fn as_item(self) -> Geometry<T> {
         Geometry::MultiPoint(self)
@@ -33,7 +32,7 @@ where
 
 impl<T> fmt::Display for MultiPoint<T>
 where
-    T: num_traits::Float + fmt::Display,
+    T: WktFloat + fmt::Display,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         if self.0.is_empty() {
@@ -54,7 +53,7 @@ where
 
 impl<T> FromTokens<T> for MultiPoint<T>
 where
-    T: num_traits::Float + FromStr + Default,
+    T: WktFloat + FromStr + Default,
 {
     fn from_tokens(tokens: &mut PeekableTokens<T>) -> Result<Self, &'static str> {
         let result =