+++ src/types/point.rs
@@ -16,15 +16,14 @@ use std::fmt;
 use std::str::FromStr;
 use tokenizer::PeekableTokens;
 use types::coord::Coord;
-use FromTokens;
-use Geometry;
+use {FromTokens, Geometry, WktFloat};
 
-#[derive(Clone, Default)]
-pub struct Point<T: num_traits::Float>(pub Option<Coord<T>>);
+#[derive(Clone, Debug, Default)]
+pub struct Point<T: WktFloat>(pub Option<Coord<T>>);
 
 impl<T> Point<T>
 where
-    T: num_traits::Float,
+    T: WktFloat,
 {
     pub fn as_item(self) -> Geometry<T> {
         Geometry::Point(self)
@@ -33,7 +32,7 @@ where
 
 impl<T> fmt::Display for Point<T>
 where
-    T: num_traits::Float + fmt::Display,
+    T: WktFloat + fmt::Display,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         match self.0 {
@@ -58,7 +57,7 @@ where
 
 impl<T> FromTokens<T> for Point<T>
 where
-    T: num_traits::Float + FromStr + Default,
+    T: WktFloat + FromStr + Default,
 {
     fn from_tokens(tokens: &mut PeekableTokens<T>) -> Result<Self, &'static str> {
         let result = <Coord<T> as FromTokens<T>>::from_tokens(tokens);