+++ CHANGES.md
@@ -5,6 +5,8 @@
 * Add `geo-types::Geometry::from(wkt)`
 * BREAKING: update geo-types, apply new `geo_types::CoordFloat`
   * <https://github.com/georust/wkt/pull/53>
+* BREAKING: Add Debug to Wkt structs by using new WktFloat instead of num_traits::Float
+  * <https://github.com/georust/wkt/pull/54>
 
 ## 0.8.0